finally enabled ssl now that we have cookies and passwords

+++ nginx/minify.sh
@@ -1,4 +1,5 @@
 #!/bin/sh
+#
 # usage: minify.sh path/to/index.html
 
 index="${1?index file must be specified}"
@@ -19,24 +20,24 @@ index="${index##*/}"
 declare -a jsfiles
 declare -a cssfiles
 
-sed -n '/<!-- begin:minify.js -->/,/<!-- end:minify.js -->/ { 
-  /<!--/ d; { 
-    s/^.*src="\.\/js\///
-    s/".*$//
-    p
-  }
-}' "$index" \
-| if cd ./js 2>/dev/null; then 
-    while read file; do 
+ if cd ./js 2>/dev/null; then 
+  sed -n '/<!-- begin:minify.js -->/,/<!-- end:minify.js -->/ { 
+    /<!--/ d; { 
+      s/^.*src="\.\/js\///
+      s/".*$//
+      p
+    }
+  }' "../$index" \
+  | while read file; do 
       cat "$file"
       echo ';' # just in case
       test -n "$MIN_CLEAN" && rm "$file"
     done >cffc-cat.js
     cd ..
-  else
-    echo "couldn't find a 'js/' directory under index-root: '`pwd`'"
-    echo "skipping js minification"
-  fi 
+else
+  echo "couldn't find a 'js/' directory under index-root: '`pwd`'"
+  echo "skipping js minification"
+fi 
 
 if test "$?" -ne "0"; then
   echo "failed to parse javascript header section from ${index}" >&2
@@ -50,31 +51,31 @@ elif ! cat >temp.html <<-EOF
   `sed -n '/end:minify.js/,$ p' "$index"`
 EOF
 then
-  echo "couldn't create temporary index" >&2
+  echo "couldn't create temporary index for js" >&2
   exit 1
 elif mv -v temp.html "$index"; then
   rm ./js/cffc-cat.js
 else
-  echo "couldn't replace '$index'" >&2
+  echo "couldn't replace '$index' for js" >&2
 fi
 
-sed -n '/<!-- begin:minify.css -->/,/<!-- end:minify.css -->/ { 
-  /<!--/ d; { 
-    s/^.*href="\.\/css\///
-    s/".*$//
-    p
-  }
-}' "$index" \
-| if cd ./css 2>/dev/null; then 
-    while read file; do 
+if cd ./css 2>/dev/null; then 
+  sed -n '/<!-- begin:minify.css -->/,/<!-- end:minify.css -->/ { 
+    /<!--/ d; { 
+      s/^.*href="\.\/css\///
+      s/".*$//
+      p
+    }
+  }' "../$index" \
+  | while read file; do 
       cat "$file"
       test -n "$MIN_CLEAN" && rm "$file" >&2
     done >cffc-cat.css
     cd .. 2>/dev/null
-  else
-    echo "couldn't find a 'css/' directory under index-root: '`pwd`'"
-    echo "skipping css minification"
-  fi 
+else
+  echo "couldn't find a 'css/' directory under index-root: '`pwd`'"
+  echo "skipping css minification"
+fi 
 
 if test "$?" -ne "0"; then
   echo "failed to parse css header section from ${index}" >&2
@@ -88,12 +89,11 @@ elif ! cat >temp.html <<-EOF
   `sed -n '/end:minify.css/,$ p' "$index"`
 EOF
 then
-  echo "couldn't create temporary index" >&2
+  echo "couldn't create temporary index for css" >&2
   exit 1
 elif mv -v temp.html "$index"; then
   echo "placeholder"
   # rm ./css/cffc-cat.css # FIXME: till minifycss above works
 else
-  echo "couldn't replace '$index'" >&2
+  echo "couldn't replace '$index' for css" >&2
 fi
-
